rename services to match ART style guides
