+++ modules/accredited_representative_portal/spec/services/power_of_attorney_request_service/create/form_data_adapter_spec.rb
@@ -2,7 +2,7 @@
 
 require 'rails_helper'
 
-RSpec.describe PowerOfAttorneyRequestCreation::FormDataAdapter do
+RSpec.describe PowerOfAttorneyRequestService::Create::FormDataAdapter do
   describe '#call' do
     subject { described_class.new(data: data, dependent: dependent, service_branch: service_branch) }
